+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## v0.0.29
+
+* timeline process configured with paths for ports
+
+## v0.0.28
+
+* vivarium/plots directory
+* register in module init
+* get_initialize_state for Generator
+* generalize timeline process
+
 ## v0.0.20
 
 * Update documentation with: